+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 2020-08-24
 
 * Deck.Date can now be null if this information was not present on the MTGO website, previously it was doing an automated fallback to Tournament.Date.
+
+2020-11-24
+
+* Added Bracket property
+* Added Standings property